requirements.txt
appdirs==1.4.3
certifi==2017.4.17
future==0.16.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-telegram-bot==5.3.0
six==1.10.0
urllib3==1.20
